we bump up the health checks from Nginx Plus default
default is 5s and we bump that to 10s
fails parameter requires the server to fail three health checks to be marked as unhealthy (up from the default one).
passes parameter means the server must pass two consecutive checks to be marked as healthy again instead of the default one.