[bibdata] reload daemons before restarting sqs poller
This should ensure systemd picks up changes to the file before
restarting the service

closes #2427

Note I think this should work but it didn't when I ran it, possibly
because I had already replaced the file on the server in a previous ansible run.
I ran this ad-hoc from the ansible command line and that did work.