add a pre_task to check for a users
There is a pre_task that checks to see if our project user exists.
This is the user that will manage our ec2 instance. If the user does not
exist a new one will be created. Otherwise we will use the users keys to
create a new VM.
This allows us to restrict what ansible can do on AWS

Co-authored-by: Alicia Cozine<acozine@users.noreply.github.com>