[Catalog] increase burst for catalog rate limit (#4074)
When a human user visits the catalog using a web browser,
it causes a burst of ~16 requests for all the images,
javascript, etc.  To reduce the chance that bursts like
this would cause nginx to block a human user, let's
increase the burst setting, but keep the rate limit
in place to block traffic that is consistently making
many requests/second.