[Postgresql] Move postgresql common role to playbook. (#4282)
* [postgresql] Move postgresql common role to playbook.

It's both a "set a server up" playbook and a "apps include this"
playbook - for "set a server up" it should be in the playbook, for "apps
include this" common not being a dependency will allow us to optimize
setting up required databases.

* Tag postgresql

By tagging postgresql we can run playbooks and just have it set up
access/databases/users. This is useful for migrations.