Remove bundler version management
Modern versions of bundler comply with gemfile configuration by checking
the "bundled with" line and downloading the desired bundler version
before executing bundler. (if it's already installed it'll just use the
correct installed version). This means we don't need to manage bundler
versions carefully like we did before.

Also, this specific bundler version management causes problems in the
context of the changes introduced in
https://github.com/pulibrary/princeton_ansible/pull/4498 -- when
passenger preloads bundler, ruby looks for a version of bundler that
appears to be locked in at installation time when installed from source.
I couldn't find a way to configure ruby to use a specific bundler
version during installation from source. So we have to ensure that the
version that ruby will look for is installed.