[researchdata] use nginx role (#5141)
* add nginx role

we are decoupling the nginxplus from nginx upstream. We are unable to
continue using the nginxplus role to run some of our PHP apps. This role
allows a project to use upstream (not the Ubuntu package one which can
lag behind)

* make the research data role use nginx

we are unable to install researchdata using nginxplus. We add to new VMs
to allow A/B testing and make the role use the new nginx role

* Update roles/nfsserver/tasks/main.yml

Co-authored-by: Max Kadel <mkadel@princeton.edu>

* include nginx role in tests

lets add nginx to the list of roles that is tested on CI

* our containers do not use systemd

* downgrade docker driver

* use requests 2.31

* revert to prancible image

---------

Co-authored-by: Max Kadel <mkadel@princeton.edu>