[video_reserves] Allow all traffic to videoreserves-prod.princeton.edu, not just princeton network
Previously, we only allowed requests to videoreserves-prod.princeton.edu if you were on
the Princeton network.  Other users needed to access the site via a library.princeton.edu
proxy pass rule.

We are going to get rid of that proxy pass rule as part of the website migration, so
we're telling the load balancer to allow traffic to this domain.