[orangelight-staging] send compressed assets if available and requested by the user's browser
The catalog application servers are already able to handle requests for gzipped assets,
which is great, because that can make things much faster for users.

However, before this commit, the load balancer was causing us to lose this feature.
Fortunately, telling the load balancer to use HTTP 1.1 (instead of the default 1.0)
allows the content negotiation to still happen.

Helps with pulibrary/orangelight#4054