[Nomad] Add host volumes for Loki + Grafana
For staging we need a persistent storage location for Loki + Grafana. We
can accomplish this most easily by just picking a box to store that data
on, and telling Nomad about it. We're not concerned about migration
right now, since this is just a staging set up.

Documentation on host volumes:
https://developer.hashicorp.com/nomad/tutorials/stateful-workloads/stateful-workloads-host-volumes