[orangelight] Make it harder for the load balancer healthcheck to take the catalog down
We observed a pattern:
1) While we had heavy bot traffic, the healthcheck would fail or timeout when the load balancer contacted it
2) In the nginx dashboard, we saw that these healthcecks would cause the VM to be taken out of the rotation.
3) The two remaining application servers had to deal with an increased amount of bot traffic with their friend down.
4) Their healthcheck started failing too.

This PR only partially helps with this issue, by increasing the number of healthchecks
required before taking a VM out of rotation, and reducing the number of healthchecks
required before adding it back.

A better fix would be to increase the amount of time before a request times out, and
adding more application servers to handle the load.

Co-authored-by: Christina Chortaria <christinach@users.noreply.github.com>
Co-authored-by: Ryan Laddusaw <rladdusaw@users.noreply.github.com>