[orangelight] increase the passenger max pool size
This increases the number of simultaneous requests that passenger will work on.
The tradeoff is that more simultaneous requests will require more memory to
process.

According to the passenger documentation, we could safely process about 28
simultaneous requests (rather than our current 6) with the current memory
usage patterns and memory available on the boxes:
https://www.phusionpassenger.com/library/config/nginx/optimization/#tuning-the-application-process-and-thread-count

Perhaps this could make the catalog more resilient to large spikes in
bot traffic?