[allsearch api] compile ruby with the YJIT in staging
Recent versions of Ruby 3.3 will automatically compile the yjit if they
notice that the rust compiler is installed.

Rails 7.2 (which Allsearch API uses) automatically uses the YJIT if it
is available.

Trying this out to see if it boosts performance at all (which we can
monitor in the datadog APM).

One way to check if the YJIT is installed is with the following shell command:

ruby -e 'puts "YJIT is#{defined?(RubyVM::YJIT) ? "" : " not"} installed"'