[fail2ban/nginxplus] Change format of fail2ban match (#5659)
Previously, we were trying to match CLF formatted logs in access.log.  But our access log
is in a JSON format, so we needed to adjust the fail2ban failregex accordingl[fail2ban/nginxplus] Change format of fail2ban match

Previously, we were trying to match CLF formatted logs in access.log.  But our access log
is in a JSON format, so we needed to adjust the fail2ban failregex accordingly

Co-authored-by: Alicia Cozine <acozine@users.noreply.github.com>
Co-authored-by: Christina Chortaria <christinach@users.noreply.github.com>
Co-authored-by: Francis Kayiwa <kayiwa@users.noreply.github.com>
Co-authored-by: James R. Griffin III <jrgriffiniii@users.noreply.github.com>
Co-authored-by: Kevin Reiss <kevinreiss@users.noreply.github.com>