Create a Traefik Bot Wall for LAE Staging
This is most of what's necessary to create a captcha for all catalog
paths for LAE Staging.

In this setup a request goes from Load Balancer -> Traefik -> LAE

Traefik's plugin (thank you Joe Corall) determines if a given path
(/catalog) has been visited a given number of times (2) by a given IP
range and if so, it redirects the user to a Cloudflare Turnstile. If
they pass the turnstile, it redirects them back. Feedback from
colleagues at other institutions have shown us that bots can not pass
this turnstile.

Pros:

* No code changes for our Blacklight applications.
* Very configurable.
* As resilient as Nomad - if one node goes down, another will quickly
  replace it automatically.
* We can delay our decision about paying Cloudflare for enterprise
  service.
* ???

Cons:

* We'd have to put the servers we're load balancing across into this
  Traefik configuration.
* We gotta add some deployment mechanism into princeton_ansible, maybe a
  sub-directory for Nomad stuff?
* ???

Open Questions:

1. Should we run the Bot Wall on its own machine, or are we cool using
   Nomad for this?
2. Should we have a bot wall per application, or feed all our Blacklight
   apps through this?
3. Would it be easier to just implement the gem Jrochkind's working on?
4. Do we need a customized cloudflare turnstile page per
   application, or can we make one generic enough for every app we might
   use this on?

To try this:

On an IP you haven't tried it on yet go to
https://lae-staging.princeton.edu (on VPN), and browse around in the search.