Create a Traefik Bot Wall for LAE Staging
This is most of what's necessary to create a captcha for all catalog
paths for LAE Staging.

In this setup a request goes from Load Balancer -> Traefik -> LAE

Traefik's plugin (thank you Joe Corall) determines if a given path
(/catalog) has been visited a given number of times (2) by a given IP
range and if so, it redirects the user to a Cloudflare Turnstile. If
they pass the turnstile, it redirects them back. Feedback from
colleagues at other institutions have shown us that bots can not pass
this turnstile.

Pros:

* No code changes for our Blacklight applications.
* Very configurable.
* As resilient as Nomad - if one node goes down, another will quickly
  replace it automatically.
* We can delay our decision about paying Cloudflare for enterprise
  service.
* ???

Cons:

* We'd have to put the servers we're load balancing across into this
  Traefik configuration.