create new solr9cloud role
we create a new separate solr9cloud role. We have re-written solrcloud
but these commits make it difficult to work with our current running
solr8cloud. To simplify movement forward we elect to separate these
roles
This role is capable of installing solr of any version in a greenfield
ensemble