I5740 solr backup separate (#6163)
* create new solr9cloud role

we create a new separate solr9cloud role. We have re-written solrcloud
but these commits make it difficult to work with our current running
solr8cloud. To simplify movement forward we elect to separate these
roles
This role is capable of installing solr of any version in a greenfield
ensemble

* add solr smb credentials

* create correct package

* restart daemon on servers and not CI

* make the role idempotent

* add verification

* add a bind-mount for backup

Solr 9 needs to be able to write under the solr data home and cannot
write to the mounted path.

this PR https://github.com/pulibrary/pul_solr/pull/472 will be changing
where rake task looks at. A bind-mount is a true mount point that will
allow solr8 to continue working with the new path.