[common] don't send vector's demo logs to the syslog
They take up a non-trivial amount of disk space.  Eventually,
we will want to send real logs to pulemetry, but it is not
quite ready yet.  This configuration is taken from the basic
config on the boxes, except that it uses a "file" sink to send
the entries to /dev/null