make sure the file permissions are deploy owned (#6333)
the variables for logrotate are made consistent with the common role but we prefix the name of the role and call them within the role and the group vars. we may need to test at a later date if the variables in the `group_vars/lib_jobs/common.yml` are even needed.

End to end test performed with

Co-authored-by: Angel Ruiz <aruiz1789@users.noreply.github.com>
Co-authored-by: Vickie Karasic <vickiekarasic@users.noreply.github.com>