make it possible for apps to opt in
we do not need all the production applications automatically added to the sandbox.
create a dictionary of env -> server-group names
keys will be passed using `runtime_env` and the values are the inventory group names that contain our nomad servers.
the resulting `nomad_effective_env` will then default to the cluster.