[solr9] Install plugin jars to /opt/solr/lib and /opt/solr/contrib/analysis-extras/lib (#6473)
* [solr9] Install plugin jars to /opt/solr/lib

In our solr 8 playbook, these jars are in two different directories:
* /solr/data/lib
* /opt/solr/contrib/analysis-extras/lib

Rather than recreating this, this commit standardizes them all to
/opt/solr/lib.

This commit also replaces several corrupted JARs in the solr9cloud
role with actual binaries from the solrcloud role directory

We also remove old JARs:
* CJKFoldingFilter is the old version of CJKFilterUtils.  We don't
  need it in solr 9 (it is not compatible), so let's remove it so
  we don't get confused about it.
* We don't need a copy of the old, incompatible version of the
  umich filter either.

And add new ones that we need:
* icu4j: The CJKFilterUtils depend on the class com.ibm.icu.text.Transliterator,
which is provided by icu4j.
* lucene-analysis-icu (note that a more recent version is available,
  but it is not compatible with our relatively old version of Solr 9)
* solr-analysis-extras

Closes #6471

* [solr9] Add certain JARs to the analysis-extras/lib directory

Some CDH solr configs expect certain JARs to be available at the path
/opt/solr/modules/analysis-extras/lib.

For example, see:
https://github.com/Princeton-CDH/mep-django/blob/239fec94c2d1dbbb6eefa03451073302fe6c3a55/solr_conf/conf/solrconfig.xml#L81-L83

This commit moves the necessary JARs to that path, and also copies
some JARs that were present on the production box in that path,
but were not yet managed in this solr role.