[Figgy + Postgresql] Set keepalive timeout lower.
We were having postgres timeouts which was resulting in some items not
being able to be ingested in Figgy. This value allowed those items to
ingest. The relevant honeybadger error:
https://app.honeybadger.io/projects/53391/faults/119695976

Closes https://github.com/pulibrary/figgy/issues/6800
Related to https://github.com/pulibrary/princeton_ansible/issues/2305

This will only change the setting for Figgy, and keep the default for
the other postgres cluster.